Refactoring buy now button
