+++ src/presentation/components/BuyNowButton/index.ts
@@ -1,1 +1,2 @@
+export { BuyNowWallet } from './BuyNowWallet';
 export { BuyNowBigButton } from './BuyNowBigButton';
